Add Classic
Realized I'd rather not make a bunch of different functions to call based on the type of shape, but that it's better if the shape has references to functions that implement the same behavior with variations depending on the shape.

So instead of "if poly then project_poly() elseif circle then project_circle" I can just do `shape:project(vector)` and it works.

Classes are only needed for method over-riding. AND! The user can implement their own geometry with minimal definition if they extend a polygon class and override the constructor.